requirements.txt
Django
djangorestframework
django-extensions
lndhub
django-cors-headers
django-taggit
